Refactoring: Move CV features into `cv` submodule
