+++ pikuli/_helpers.py
@@ -2,8 +2,7 @@
 
 import os
 
-from pikuli import logger
-
+from . import logger
 
 class NotImplemetedDummyBase(object):
     err_msg = None
