+++ pikuli/cv/screen.py
@@ -1,9 +1,10 @@
 # -*- coding: utf-8 -*-
 
-from . import _functions, FailExit, geom
+from .. import FailExit
+from .._functions import _screen_n_to_mon_descript
+from ..geom import Region
 
-
-class Screen(geom.region.Region):
+class Screen:
     """ Represents physical computer display screen.
         x, y  --  left upper corner coords relative to virtual desktop.
         w, h  --  screen rectangle area dimensions.
@@ -19,14 +20,18 @@ class Screen(geom.region.Region):
         # For each monitor found, returns a handle to the monitor,
         # device context handle, and intersection rectangle:
         # (hMonitor, hdcMonitor, PyRECT).
-        self.__mon_hndl, _, mon_rect = \
-            _functions._screen_n_to_mon_descript(self.n)
+        self.__mon_hndl, _, mon_rect = _screen_n_to_mon_descript(self.n)
 
-        super(Screen, self).__init__(mon_rect[0],
-                                     mon_rect[1],
-                                     mon_rect[2] - mon_rect[0],
-                                     mon_rect[3] - mon_rect[1],
-                                     title='Screen ({})'.format(self.n))
+        self._region = Region(
+            mon_rect[0],
+            mon_rect[1],
+            mon_rect[2] - mon_rect[0],
+            mon_rect[3] - mon_rect[1],
+            title='Screen ({})'.format(self.n))
+
+    @property
+    def region(self):
+        return self._region
 
     def __repr__(self):
         return '<Screen ({}) ({}, {}, {}, {})>'.format(
