+++ pikuli/geom/__init__.py
@@ -1,5 +1,7 @@
 # -*- coding: utf-8 -*-
 
+from .simple_types import Rectangle, Point
+
 from .vector import Vector
 from .location import Location
 from .region import Region
