+++ pikuli/uia/control_wrappers/check_box.py
@@ -1,7 +1,6 @@
 # -*- coding: utf-8 -*-
 
-
-from pikuli import wait_while_not, wait_while
+from pikuli.utils import wait_while_not, wait_while
 from pikuli.uia.adapter import Enums, STATE_SYSTEM
 from . import CONTROL_CHECK_TIMEOUT
 from .uia_control import UIAControl
