+++ pikuli/uia/control_wrappers/combo_box.py
@@ -1,6 +1,7 @@
 # -*- coding: utf-8 -*-
 
-from pikuli import FindFailed, wait_while_not
+from pikuli import FindFailed
+from pikuli.utils import wait_while_not
 
 from . import CONTROL_CHECK_TIMEOUT
 from .uia_control import UIAControl
